requirements.txt
Django==3.1.4
requests==2.25.1
gunicorn==20.0.4
pymysql==1.0.2
requests==2.25.1
djangorestframework==3.12.2
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
chinesecalendar==1.5.0

